fix authentication check for connection requests
zstream_b64decode does not return a null terminated buffer, but the
code assumes that it is. This results in buffer overflows until the
next null byte is encountered in the length calculation, with
possible false rejections.

Since zstream_b64decode puts the size of the buffer into size, we
can directly use that instead of calculating the string length.

Contributed by T-Labs, Deutsche Telekom Innovation Laboratories

Signed-off-by: Jonas Gorski <jonas.gorski@gmail.com>